Made drawPoint and drawLine functions compatible with cglm vec3 types. (This was done in a very hacky way but hey it works.)
Made the FPS tracking function more useful.

Separated the draw functions, FPS function, and callback functions into separate source files.
This greatly cut down on the bloat in the GPbase.c file.

Screen has 640 red lines drawn to it.
