+++ include/GPbase.h
@@ -13,9 +13,17 @@
 // Window creation library. Also handles other things like input.
 #include "GLFW/glfw3.h"
 
+// C compatible math library for OpenGL
 #include "cglm/cglm.h"
 
+// Load shaders sourced from the OpenGL Red Book
 #include "LoadShaders.h"
+
+#include "DrawFuncs.h"
+
+#include "FPSProfiling.h"
+
+#include "CallbackFunctions.h"
 
 // Custom defines
 // Global variables
